requirements.txt
flask==1.0.3
flask_mwoauth
qrcode
pyyaml
